Put library monitor on thread with ascyrio
